Moving relational operations to separate trait
