+++ src/Libraries/Database/Database.php
@@ -9,7 +9,7 @@
  * @author Arman Ag. <arman.ag@softberg.org>
  * @copyright Copyright (c) 2018 Softberg LLC (https://softberg.org)
  * @link http://quantum.softberg.org/
- * @since 2.9.5
+ * @since 2.9.6
  */
 
 namespace Quantum\Libraries\Database;
@@ -17,6 +17,7 @@ namespace Quantum\Libraries\Database;
 use Quantum\Libraries\Database\Exceptions\DatabaseException;
 use Quantum\Libraries\Config\Exceptions\ConfigException;
 use Quantum\Libraries\Database\Contracts\DbalInterface;
+use Quantum\Libraries\Database\Traits\RelationalTrait;
 use Quantum\Di\Exceptions\DiException;
 use Quantum\Loader\Setup;
 use ReflectionException;
@@ -27,6 +28,8 @@ use ReflectionException;
  */
 class Database
 {
+
+    use RelationalTrait;
 
     /**
      * Database configurations
@@ -106,99 +109,5 @@ class Database
         }
 
         return config()->get('database.' . $currentKey);
-    }
-
-    /**
-     * Raw execute
-     * @param string $query
-     * @param array $parameters
-     * @return bool
-     * @throws DatabaseException
-     */
-    public static function execute(string $query, array $parameters = []): bool
-    {
-        return self::resolveQuery(__FUNCTION__, $query, $parameters);
-    }
-
-    /**
-     * Raw query
-     * @param string $query
-     * @param array $parameters
-     * @return array
-     * @throws DatabaseException
-     */
-    public static function query(string $query, array $parameters = []): array
-    {
-        return self::resolveQuery(__FUNCTION__, $query, $parameters);
-    }
-
-    /**
-     * Fetches table columns
-     * @param string $query
-     * @param array $parameters
-     * @return array
-     * @throws DatabaseException
-     */
-    public static function fetchColumns(string $query, array $parameters = []): array
-    {
-        return self::resolveQuery(__FUNCTION__, $query, $parameters);
-    }
-
-    /**
-     * Gets the last query executed
-     * @return string|null
-     * @throws DatabaseException
-     */
-    public static function lastQuery(): ?string
-    {
-        return self::resolveQuery(__FUNCTION__);
-    }
-
-    /**
-     * Get an array containing all the queries
-     * run on a specified connection up to now.
-     * @return array
-     * @throws DatabaseException
-     */
-    public static function queryLog(): array
-    {
-        return self::resolveQuery(__FUNCTION__);
-    }
-
-    /**
-     * Gets the ORM class
-     * @return string
-     * @throws DatabaseException
-     */
-    protected function getOrmClass(): string
-    {
-        $ormClass = $this->configs['orm'];
-
-        if (!class_exists($ormClass)) {
-            throw DatabaseException::ormClassNotFound($ormClass);
-        }
-
-        if (!$ormClass::getConnection()) {
-            $ormClass::connect($this->configs);
-        }
-
-        return $ormClass;
-    }
-
-    /**
-     * Resolves the requested query
-     * @param string $method
-     * @param string $query
-     * @param array $parameters
-     * @return mixed
-     * @throws DatabaseException
-     */
-    protected static function resolveQuery(string $method, string $query = '', array $parameters = [])
-    {
-        $self = self::getInstance();
-
-        $ormClass = $self->getOrmClass();
-
-        return $ormClass::$method($query, $parameters);
     }
 }
